Refactor Sidebar into subcomponents (StylePanel, StylePreview)
